// src/components/calculatorButtons/enums.ts
/* eslint-disable no-shadow */
export enum CalcButtonType {
    operation,
    function,
    number,
    history
}

export enum OperationType {
    devide = '÷',
    miltiply = 'x',
    sum = '+',
    substraction = '-',
    equal = '='
}

export enum CalculatorFunctionType {
    clear = 'AC',
    pm = '±',
    percent = '%',
    float = '.'
}